Using pkg/errors, rewrite gameover handling
pkg/errors also stores the stack trace, and allows for
easier combining of errors.

GameOver handling was somewhat changed, as using the GameOverErr
constant did not mix well with the pkg/errors package. Instead we
allow the gameLoop to finish with a nil error to indicate a successful
run of the game. Never returning a nil error makes sense for
components that are not supposed to quit, such as the coordinator,
but the game loop is meant to finish. As such, it seems to make sense
to return nil there.