Refactors entry point, adds command line flags
Moved the http server creation into the main.go file, as to have access
to the flags passed in by the user. It also makes it easier for the user
of lib to configure where and how the http server is hosted.

Added flags for enabling debug logging (default false), serving of
static assets (default false) and setting the address where the http
server is listening (default :8080).