Sending events no longer awaits send confirmiation
Events are now queued on a new go routine per event. This essentially
means that a slower client should no longer make faster clients lag.
It also means that there might be a queue of events built up between
a client leaving and the read timeout indicating that it left.

Current read timeout of 10 sec, means we should have about ~20 tick
events at most in the queue. Could potentially add a timeout to
context in the handleEvent method in controller, if this turns out
to be a problem.