Airplanes now predict and chase server position
Instead of updating the airplane positions to the exact server
position each tick, we instead modify the local speed and direction
to reach the position where we think the airplane will be next tick.

This means we can smooth out the compensation during multiple frames,
rather than sometimes abruptly moving the airplane.