Controller no longer care for num players
The game controller now accepts any number of player proxies (> 0),
leaving pairing of the clients to the model player to its caller.

In addition, this commit also changes the controller's action reading
loop so that it continues even after a player leaves, instead of
forcefully ending the game. Doing this means we leave the logic of
determining when a game is over to the game loop.

These changes do not affect how the game plays currently, as the 2
player limit is enforced also in other places. However, it makes way for
future >2 player games.