Send tick event also for last tick
This commit refactors (primarily) the game controller and game loop to
enable sending a tick event also for the tick that resulted in game over
being sent. In order to make this possible the game loop and game
controller now guarantees in-order sending of events. I.e. each event
produced by the game loop is now put on a queue and processed in order
on a single goroutine per player.

To protect against slow receivers, each player is given a separate queue
of events to send. Should a player be slow enough that their queue is
filled up they are now disconnected. For this to work the disconnect
method of the network client was rewritten to be non-blocking, making it
safe for the single eventLoop to call disconnect without blocking the
sending of events to other players.

In addition to sending event in-order, the game controller now also
waits for pending events before quitting. This is realized by separating
forceful closure of the game controller (via its context) from normal
shutdown.