web, wire: Remove 2 player limit
This commit changes the wire proto to encode players as a repeated field
and handles this new field on the client side.

Due to changes in wire protocol (and no need to be backwards compatible)
the version was bumped to v3.

**Note**: The .js and .ts files were generated using a locally modified
protobufjs that used 3.6.2 instead, which seems to be required to
generate non-broken .ts files.