Allow games with up to 10 players
This commit makes the game coordinator wait for up to 10 players before
starting a game. Once the minimum number of players is reached (=2) a
timeout is started. If no new players connect within the timeout then
the game starts. If a new player connects then the timeout is reset. If
the number of connected players reached the max number of players (=10)
then the game is immediately started.