web: distinct colors for multiple players
This commit changes from using a hard-coded red color for all non-self
non-neutral players, to use a specific color per player.

The color per player is decided entirely client-side based on index in
the players array of the tick event. This should be deterministic among
all clients, but is probably not an ideal solution for the future. The
implementation was made with an assumption that the player model
communicated by the server will eventually contain a color selected by
the player themselves.

The "self" and "neutral" players are still blue and white respectively.

Finally, this commit makes the island ownership color more visible by
switching from coloring the text to tinting the entire island instead.
The island assets were made black/white to make this tinting possible.