Introduce exclusive player states
Instead of the player having a bunch of getters for various states that
was implicitly exclusive, this commit adds a PlayerState to each Player.

This commit also moves checking of player alive to the tick action. The
check for game over is still performed in the check game over action.

This commit makes it fairly obvious that the tick action is a very
central part of the game and should probably be rethought in the future.