Add AI opponents
The AI opponents are implemented as Clients, functioning in the same way
as remote clients but short-circuiting the need to transfer events and
actions over the wire. Currently the AI Client performs zero or one
action per event, which means it is tied to the game loop tick interval.

The behavior of a Client is controlled by a "strategy". There are
currently two strategies, one that simply sends at random, and one that
tries to be a bit more clever. Neither strategy has any understanding of
past event.