Wire: Assume missing speed is default
Most airplanes will have the default speed, so we can reduce the bytes
sent by leaving it out on the server and assuming a missing value means
the default speed on the client.

Note that this prevents us from having a speed of 0 (since it will be
interpreted as the default speed), but that seems like something we will
currently never want anyway.