Server: Make Server an http router
This commit restructures the Server to handle routing for a new rooted
HTTP path "/app/". This makes it possible to add additional routes to
the server easily.

As part of this restructuring the websocket upgrader was simplified to
basically be a thin wrapper around gorilla upgrader. This previous
callback based approach was used to dynamically set the context used for
new clients. This was instead implemented by dynamically changing a
clientContext factory function in Server.Run.